rename templates view to wge view
